# users/openapi.yaml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
  description: API for managing users

paths:
  /user:
    post:
      summary: Creates a new user
      description: Creates a new user with the given username.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  example: "testuser"
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User testuser created successfully in MongoDB"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '409':
          description: User already exists

  /login:
    post:
      summary: User Login
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "testuser"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: Login successful
        '401':
          description: Incorrect username or password

  /user/{username}:
    get:
      summary: Gets a user profile
      description: Returns the public account data for the given user.
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
    put:
      summary: Updates a user profile
      description: Updates name, surname and email for the given user.
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid profile data
        '404':
          description: User not found

  /user/{username}/history:
    get:
      summary: Gets game history
      description: Returns the list of games played by a specific user.
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSession'
        '404':
          description: User not found

  /game/finish:
    post:
      summary: Saves a game session
      description: Records the details of a finished game in the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameSession'
      responses:
        '201':
          description: Game recorded successfully

  /logout:
    post:
      summary: Logs out a user
      description: Updates the lastLogoutAt timestamp of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  example: "testuser"
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User testuser logged out"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Username missing
        '404':
          description: User not found

components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "65cfa4b8f1a2c34d56789abc"
        username:
          type: string
          example: "testuser"
        createdAt:
          type: string
          format: date-time
          example: "2026-02-24T18:25:43.511Z"
        lastLogoutAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-02-24T19:00:12.321Z"
    UserProfile:
      type: object
      properties:
        _id:
          type: string
          example: "65cfa4b8f1a2c34d56789abc"
        username:
          type: string
          example: "testuser"
        name:
          type: string
          example: "Mario"
        surname:
          type: string
          example: "Trelles"
        email:
          type: string
          example: "mario@uniovi.es"
    GameSession:
      type: object
      required:
        - userId
        - rival
        - level
        - duration
        - result
      properties:
        userId:
          type: string
          example: "testuser"
        rival:
          type: string
          example: "bot"
        level:
          type: string
          example: "Medium"
        duration:
          type: integer
          description: Duration in seconds
          example: 120
        result:
          type: string
          example: "won"
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      required:
        - name
        - surname
        - email
      properties:
        name:
          type: string
          example: "Mario"
        surname:
          type: string
          example: "Trelles"
        email:
          type: string
          example: "mario@uniovi.es"